requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
neo4j==5.14.1
pandas==2.1.3
numpy==1.26.2
torch==2.1.1
torch-geometric==2.4.0
scikit-learn==1.3.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
pydantic==2.5.2
langchain==0.0.350
chromadb==0.4.22 